# public/api-docs.yaml
openapi: 3.0.0
info:
  title: ScamGuard API
  description: Comprehensive scam detection and prevention API
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
  /api/categories:
    get:
      tags: [Categories]
      summary: Get categories
      responses:
        200:
          description: OK
    post:
      tags: [Categories]
      summary: Fetch category data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category]
              properties:
                category: {type: string}
                page: {type: integer}
      responses:
        200:
          description: OK
...
  /api/detail-feedback:
    get:
      tags: [Feedback]
      summary: Get feedback
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: id
          in: query
          schema:
            type: string
      responses:
        200:
          description: OK
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT